# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: providerId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review for a completed request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List all skills
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [built_in, custom, all]
          description: Filter by skill type
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, all]
          description: Filter by status (defaults to approved)
      responses:
        "200":
          description: Skills list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [skills]
      summary: Create a new skill (custom)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Skill already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /skills/{id}:
    get:
      operationId: getSkill
      tags: [skills]
      summary: Get a skill by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateSkill
      tags: [skills]
      summary: Update a skill (admin — review/approve a proposed service)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSkillInput"
      responses:
        "200":
          description: Updated skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Delete a skill (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /provider-skills:
    get:
      operationId: listProviderSkills
      tags: [skills]
      summary: List skills for the current provider
      responses:
        "200":
          description: Provider skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderSkill"
        "404":
          description: No provider profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: setProviderSkills
      tags: [skills]
      summary: Set skills for the current provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetProviderSkillsInput"
      responses:
        "200":
          description: Updated skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderSkill"
        "404":
          description: No provider profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Review:
      type: object
      required: [id, requestId, providerId, userId, rating, createdAt]
      properties:
        id:
          type: integer
        requestId:
          type: integer
        providerId:
          type: integer
        userId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ReviewInput:
      type: object
      required: [requestId, providerId, rating]
      properties:
        requestId:
          type: integer
        providerId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
...
    Skill:
      type: object
      required: [id, name, slug, type, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameEn:
          type: ["string", "null"]
        nameHe:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        descriptionEn:
          type: ["string", "null"]
        descriptionHe:
          type: ["string", "null"]
        slug:
          type: string
        category:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        type:
          type: string
          enum: [built_in, custom]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdByUserId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    SkillInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        category:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [built_in, custom]
          default: custom
    UpdateSkillInput:
      type: object
      description: Admin edit of a skill (review/approve a proposed service).
      properties:
        name:
          type: string
          minLength: 1
        nameEn:
          type: string
        nameHe:
          type: string
        description:
          type: string
        descriptionEn:
          type: string
        descriptionHe:
          type: string
        category:
          type: string
        icon:
          type: string
        color:
          type: string
        image:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
    ProviderSkill:
      type: object
      required: [id, providerId, skillId, createdAt]
      properties:
        id:
          type: integer
        providerId:
          type: integer
        skillId:
          type: integer
        experienceYears:
          type: ["integer", "null"]
        skill:
          $ref: "#/components/schemas/Skill"
        createdAt:
          type: string
          format: date-time
    SetProviderSkillsInput:
      type: object
      required: [skillIds]
      properties:
        skillIds:
          type: array
          items:
            type: integer
          description: IDs of skills to assign to the provider